Relocate SSH test to /test folder.
This better demonstrates actual, real-world usage of Terratest with GCP.
