+++ examples/terraform-gcp-example/main.tf
@@ -1,12 +1,15 @@
+
+
 # ---------------------------------------------------------------------------------------------------------------------
 # DEPLOY A CLOUD INSTANCE RUNNING UBUNTU
 # See test/terraform_gcp_example_test.go for how to write automated tests for this code.
 # ---------------------------------------------------------------------------------------------------------------------
 
 resource "google_compute_instance" "example" {
-  name         = "${var.instance_name}"
+  project = "${var.gcp_project_id}"
+  name = "${var.instance_name}"
   machine_type = "${var.machine_type}"
-  zone         = "${var.zone}"
+  zone = "${var.zone}"
 
   boot_disk {
     initialize_params {
@@ -16,6 +19,7 @@ resource "google_compute_instance" "example" {
 
   network_interface {
     network = "default"
+    access_config {}
   }
 }
 
@@ -24,6 +28,7 @@ resource "google_compute_instance" "example" {
 # ---------------------------------------------------------------------------------------------------------------------
 
 resource "google_storage_bucket" "example_bucket" {
-  name     = "${var.bucket_name}"
+  project = "${var.gcp_project_id}"
+  name = "${var.bucket_name}"
   location = "${var.bucket_location}"
 }
