+++ modules/gcp/compute.go
@@ -514,3 +514,11 @@ func NewInstancesServiceE(t *testing.T) (*compute.InstancesService, error) {
 
 	return service.Instances, nil
 }
+
+// Return a random, valid name for GCP Compute Instances. Note that GCP requires Instance names to use lowercase letters only.
+func UniqueGcpInstanceName() string {
+	id := strings.ToLower(random.UniqueId())
+	instanceName := fmt.Sprintf("terratest-%s", id)
+
+	return instanceName
+}
